bugfix: support calls to os.getenv wrapper even after unwrapping
Fix a bug where the `os.getenv` wrapper function in misc.lua stops
working after it "unwraps" and restores the original binding of
`os.getenv`.

The wrapper function can still be referenced and called by a user and
so we need to keep it working even after the original binding of
`os.getenv` has been restored.

Consider this example:

    local getenv = os.getenv
    getenv("foo")
    getenv("bar")

which is problematic if *both* calls attempt to unwrap `os.getenv`.

Here is a real-world example that does trigger an error:
https://github.com/daurnimator/lua-http/blob/master/http/proxies.lua#L18-L32