tests: ssl.t: updated a test to pass with OpenSSL 1.1.1 CLI binaries.
This test case invokes the `openssl` binary from a shell to execute a
handshake. It is supposed to be for old versions of OpenSSL, but since
we can't skip the test with `skip_openssl` (since it isn't about the
linked OpenSSL lib), we updated the subject line check to work for
OpenSSL 1.1.1 CLI binaries.

With older versions of OpenSSL:

    subject=/C=US/ST=California/L=San Francisco/O=OpenResty/OU=OpenResty/CN=test.com/emailAddress=agentzh@gmail.com

With OpenSSL 1.1.1:

    subject=C = US, ST = California, L = San Francisco, O = OpenResty, OU = OpenResty, CN = test.com, emailAddress = agentzh@gmail.com