bugfix: now 'ssl.get_tls1_version_str()' returns the error "unknown version" when the TLS version number is not recognized.
In the future, this will avoid bugs in user applications when the `ver`
return value is expected to be non-nil, but is `nil` since the TLS
version number has not been registered in the Lua module yet.

Also includes a few grammatical and styling changes to the ssl.md
documentation.